Refactor CMake configurations and enhance game functionality
- Updated CMakeLists.txt to improve compiler warning settings and ensure compatibility across different compilers.
- Added new CMake options for better handling of warnings and sanitizers.
- Introduced new variables in game source files to prevent empty translation units when certain features are disabled.
- Implemented a minimal default menu and HUD script to streamline UI loading and prevent missing file errors.
- Enhanced the encounter system with additional Lua integration for better event handling.

These changes collectively improve the build process, enhance game mechanics, and provide better support for UI and scripting functionalities.

+++ src/client/snd_mix.c
@@ -402,7 +402,7 @@ static void S_TransferPaintBuffer( int endtime, byte *buffer )
 		out_idx = ( s_paintedtime * dma.channels ) % dma.samples;
 		while ( count > 0 ) {
 			int n = count;
-			if ( n + out_idx > dma.samples )
+			if ( (unsigned int)(n + out_idx) > dma.samples )
 				n = dma.samples - out_idx;
 			CL_WriteAVIAudioFrame( buffer + out_idx * dma.samplebits / 8, n * dma.samplebits / 8 );
 			out_idx = (out_idx + n) % dma.samples;
