+++ src/qcommon/history.c
@@ -179,14 +179,14 @@ static void Con_LoadHistory( void )
 
 			numChars = atoi( token );
 			text_p++;
-			if ( numChars > ( consoleSaveBufferSize - ( text_p - consoleSaveBuffer ) ) || numChars >= sizeof( edit->buffer ) )
+			if ( numChars > (int)( consoleSaveBufferSize - ( text_p - consoleSaveBuffer ) ) || numChars >= (int)sizeof( edit->buffer ) )
 			{
 				Com_DPrintf( S_COLOR_YELLOW "WARNING: probable corrupt history\n" );
 				break;
 			}
 
-			if ( edit->cursor > sizeof( edit->buffer ) - 1 )
-				edit->cursor = sizeof( edit->buffer ) - 1;
+			if ( edit->cursor > (int)sizeof( edit->buffer ) - 1 )
+				edit->cursor = (int)sizeof( edit->buffer ) - 1;
 			else if ( edit->cursor < 0 )
 				edit->cursor = 0;
