+++ src/renderer/tr_cmds.c
@@ -130,8 +130,8 @@ static void *R_GetCommandBufferReserved( int bytes, int reservedBytes ) {
 	bytes = PAD(bytes, sizeof(void *));
 
 	// always leave room for the end of list command
-	if ( cmdList->used + bytes + sizeof( int ) + reservedBytes > MAX_RENDER_COMMANDS ) {
-		if ( bytes > MAX_RENDER_COMMANDS - sizeof( int ) ) {
+	if ( (size_t)cmdList->used + (size_t)bytes + sizeof( int ) + (size_t)reservedBytes > (size_t)MAX_RENDER_COMMANDS ) {
+		if ( (size_t)bytes > (size_t)MAX_RENDER_COMMANDS - sizeof( int ) ) {
 			ri.Error( ERR_FATAL, "R_GetCommandBuffer: bad size %i", bytes );
 		}
 		// if we run out of room, just start dropping commands
