+++ src/renderervk/vk_vbo.c
@@ -358,13 +358,13 @@ static void VBO_AddGeometry( vbo_t *vbo, vbo_item_t *vi, shaderCommands_t *input
 
 	offs = input->shader->iboOffset + input->shader->curIndexes * sizeof( input->indexes[0] );
 	size = input->numIndexes * sizeof( input->indexes[ 0 ] );
-	if ( offs + size > vbo->vbo_size ) {
+	if ( offs + size > (uint32_t)vbo->vbo_size ) {
 		ri.Error( ERR_DROP, "Index0 overflow" );
 	}
 	memcpy( vbo->vbo_buffer + offs, input->indexes, size );
 
 	// fill soft buffer too
-	if ( vbo->ibo_offset + size > vbo->ibo_size ) {
+	if ( (uint32_t)(vbo->ibo_offset + size) > (uint32_t)vbo->ibo_size ) {
 		ri.Error( ERR_DROP, "Index1 overflow" );
 	}
 	memcpy( vbo->ibo_buffer + vbo->ibo_offset, input->indexes, size );
@@ -374,7 +374,7 @@ static void VBO_AddGeometry( vbo_t *vbo, vbo_item_t *vi, shaderCommands_t *input
 	// vertexes
 	offs = input->shader->vboOffset + input->shader->curVertexes * sizeof( input->xyz[0] );
 	size = input->numVertexes * sizeof( input->xyz[ 0 ] );
-	if ( offs + size > vbo->vbo_size ) {
+	if ( offs + size > (uint32_t)vbo->vbo_size ) {
 		ri.Error( ERR_DROP, "Vertex overflow" );
 	}
 	//Com_Printf( "v offs=%i size=%i\n", offs, size );
@@ -383,7 +383,7 @@ static void VBO_AddGeometry( vbo_t *vbo, vbo_item_t *vi, shaderCommands_t *input
 	// normals
 	offs = input->shader->normalOffset + input->shader->curVertexes * sizeof( input->normal[0] );
 	size = input->numVertexes * sizeof( input->normal[ 0 ] );
-	if ( offs + size > vbo->vbo_size ) {
+	if ( offs + size > (uint32_t)vbo->vbo_size ) {
 		ri.Error( ERR_DROP, "Normals overflow" );
 	}
 	//Com_Printf( "v offs=%i size=%i\n", offs, size );
@@ -394,7 +394,7 @@ static void VBO_AddGeometry( vbo_t *vbo, vbo_item_t *vi, shaderCommands_t *input
 	if( vk.pbrActive ) {	
 		offs = input->shader->qtangentOffset + input->shader->curVertexes * sizeof(input->qtangent[0]);
 		size = input->numVertexes * sizeof(input->qtangent[0]);
-		if (offs + size > vbo->vbo_size) {
+		if ( offs + size > (uint32_t)vbo->vbo_size) {
 			ri.Error(ERR_DROP, "Qtangent overflow");
 		}
 		//Com_Printf( "v offs=%i size=%i\n", offs, size );
@@ -403,7 +403,7 @@ static void VBO_AddGeometry( vbo_t *vbo, vbo_item_t *vi, shaderCommands_t *input
 		// lightdir
 		offs = input->shader->lightdirOffset + input->shader->curVertexes * sizeof(input->lightdir[0]);
 		size = input->numVertexes * sizeof(input->lightdir[0]);
-		if (offs + size > vbo->vbo_size) {
+		if ( offs + size > (uint32_t)vbo->vbo_size) {
 			ri.Error(ERR_DROP, "Lightdir overflow");
 		}
 		//Com_Printf( "v offs=%i size=%i\n", offs, size );
